Move add() method from api.py to add.py
