+++ dbcollection/core/api.py
@@ -22,7 +22,7 @@ from __future__ import print_function
 from .download import download
 from .process import process
 from .load import load
-from .add import AddAPI
+from .add import add
 from .remove import RemoveAPI
 from .config_cache import ConfigAPI
 from .query import QueryAPI
@@ -35,56 +35,6 @@ def check_if_dataset_name_is_valid(name):
     """Check if the dataset name exists (is valid) in the list of available dataset for download"""
     available_datasets_list = fetch_list_datasets()
     assert name in available_datasets_list, 'Invalid dataset name: {}'.format(name)
-
-
-def add(name, task, data_dir, file_path, keywords=(), verbose=True, is_test=False):
-    """Add a dataset/task to the list of available datasets for loading.
-
-    Parameters
-    ----------
-    name : str
-        Name of the dataset.
-    task : str
-        Name of the task to load.
-    data_dir : str
-        Path of the stored data in disk.
-    file_path : bool
-        Path to the metadata HDF5 file.
-    keywords : list of strings, optional
-        List of keywords to categorize the dataset.
-    verbose : bool, optional
-        Displays text information (if true).
-    is_test : bool, optional
-        Flag used for tests.
-
-    Examples
-    --------
-    Add a dataset manually to dbcollection.
-
-    >>> import dbcollection as dbc
-    >>> dbc.add('new_db', 'new_task', 'new/path/db', 'newdb.h5', ['new_category'])
-    >>> dbc.query('new_db')
-    {'new_db': {'tasks': {'new_task': 'newdb.h5'}, 'data_dir': 'new/path/db', 'keywords':
-    ['new_category']}}
-
-    """
-    assert name, "Must input a valid name: {}".format(name)
-    assert task, "Must input a valid task: {}".format(task)
-    assert data_dir, "Must input a valid data_dir: {}".format(data_dir)
-    assert file_path, "Must input a valid file_path: {}".format(file_path)
-
-    db_adder = AddAPI(name=name,
-                      task=task,
-                      data_dir=data_dir,
-                      file_path=file_path,
-                      keywords=keywords,
-                      verbose=verbose,
-                      is_test=is_test)
-
-    db_adder.run()
-
-    if verbose:
-        print('==> Dataset registry complete.')
 
 
 def remove(name, task=None, delete_data=False, verbose=True, is_test=False):
